refactor: as-is 제거 및 to-be 패키지 제거
