+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -4,7 +4,7 @@ import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.adapter.ManualHandlerAdapter;
 import nextstep.mvc.adapter.AnnotationHandlerAdapter;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
