+++ mvc/src/test/java/nextstep/mvc/controller/tobe/ControllerScannerTest.java
@@ -1,13 +1,13 @@
 package nextstep.mvc.controller.tobe;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Map;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import nextstep.mvc.controller.ControllerScanner;
 import samples.TestController;
 
 class ControllerScannerTest {
